Make sfx into a trenchbroom node.
